requirement.txt
python-docx==0.8.11
docxcompose==1.3.0
argparse==1.4.0